requirements.txt
discord.py
aiohttp
htmllaundry